Add players
Using a 'factory function' pattern, add a player
to the game. His default name is 'Player 1' and
the dafault piece is 'x'.

Add a second player, the enemy, whose default name
is 'Player 2' and the dafault piece is 'o'.

Add two buttons to represent the available pieces.
By clicking one of the buttons, the 'Player 1' swaps
pieces, and automatically updates 'Player 2'.