Move the result conditionals from 'game' to 'winCondition()'
Previously, the code was redundant, we were
calling a function named 'winCondition()'
but we also were checking player pieces and
combinations outside of the said function.

To avoid confusion, now everything about the
result, whether win or draw, is contained
in just one place.

The 'game' module is cleaner and more
readable.