Player opponent and fixes
So far, the only available opponent
for 'Player 1' is another human and it's a
local match.

If it's not clear enough, it means that you
can't play against the computer/AI. Furthermore,
it's only playable on 'player 1' machine, you can't
play against another player over the internet.

As soon as possible, I'll add more features.

Given these decisions, we are removing buttons
to pick a piece/side, since in a Tic Tac Toe
game the common rule is 'x' is the 'player 1' and
they start the match.

Restart button also affect result exhibition.